+++ eelbrain/_wxgui/tests/test_select_components.py
@@ -17,7 +17,7 @@ def test_select_components():
     data_path = mne.datasets.testing.data_path()
     raw_path = join(data_path, 'MEG', 'sample', 'sample_audvis_trunc_raw.fif')
     raw = mne.io.Raw(raw_path, preload=True)
-    ds = load.fiff.events(raw)
+    ds = load.fiff.events(raw, stim_channel='STI 014')
     ds['epochs'] = load.fiff.mne_epochs(ds, tmax=0.1)
     ica = mne.preprocessing.ICA(0.95, max_iter=1)
     with catch_warnings():