+++ eelbrain/_wxgui/tests/test_select_epochs.py
@@ -19,7 +19,7 @@ def test_select_epochs():
     data_path = mne.datasets.testing.data_path()
     raw_path = join(data_path, 'MEG', 'sample', 'sample_audvis_trunc_raw.fif')
     raw = mne.io.Raw(raw_path, preload=True).pick_types('mag', stim=True)
-    ds = load.fiff.events(raw)
+    ds = load.fiff.events(raw, stim_channel='STI 014')
     ds['meg'] = load.fiff.epochs(ds, tmax=0.1)
     # 25 cases
     arange = np.arange(25)